Extrude createHarness factory.
